Use ssize_t instead of long in mtrand

+++ numpy/random/mtrand/initarray.c
@@ -107,13 +107,13 @@ init_genrand(rk_state *self, unsigned long s)
  * key_length is its length
  */
 extern void
-init_by_array(rk_state *self, unsigned long init_key[], unsigned long key_length)
+init_by_array(rk_state *self, unsigned long init_key[], size_t key_length)
 {
     /* was signed in the original code. RDH 12/16/2002 */
-    unsigned int i = 1;
-    unsigned int j = 0;
+    size_t i = 1;
+    size_t j = 0;
     unsigned long *mt = self->key;
-    unsigned int k;
+    size_t k;
 
     init_genrand(self, 19650218UL);
     k = (RK_STATE_LEN > key_length ? RK_STATE_LEN : key_length);
